pyproject.toml
[tool.poetry]
name = "basic_django"
version = "0.1.0"
description = "Basic Django with poetry and tests"
authors = ["Arthur Antunes <arthur.antunes.aa@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
coverage = "^6.3.2"
djangorestframework = "^3.13.1"
Markdown = "^3.3.7"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
